Moving raising event to EventStatusService
