+++ src/main/application/services/ui/ChangeUiVisibleUseCase.js
@@ -1,41 +1,21 @@
-import {DomainEventBus} from '../../../domain/service/DomainEventBus'
 import {inject} from '../../../core/ioc/ioc'
 import {DisplayStatusRepository} from '../../../domain/status/DisplayStatusRepository'
 import {DisplayStatus} from '../../../domain/status/DisplayStatus'
-import {EVENT_STATUS, EventStatus} from '../../../domain/status/EventStatus'
+import {EventStatusService} from '../../../domain/service/EventStatusService'
 
 export class ChangeUiVisibleUseCase {
   constructor({
-    domainEventBus = inject(DomainEventBus),
-    displayStatusRepository = inject(DisplayStatusRepository)
+    displayStatusRepository = inject(DisplayStatusRepository),
+    eventStatusService = inject(EventStatusService)
   } = {}) {
-    this._domainEventBus = domainEventBus
     this._displayStatusRepository = displayStatusRepository
+    this._eventStatusService = eventStatusService
   }
 
   execute({visible}) {
-    let eventStatus
-    if (visible === false) {
-      this._displayStatusRepository.setDisplayStatus({
-        newStatus: DisplayStatus.HIDDEN
-      })
-      eventStatus = EventStatus.USERACTIONCOMPLETE
-    } else {
-      this._displayStatusRepository.setDisplayStatus({
-        newStatus: DisplayStatus.VISIBLE
-      })
-      eventStatus = EventStatus.CMPUISHOWN
-    }
-    // TODO  get TCData
-    const TCData = {
-      description: 'This TCData should be get from repository',
-      eventStatus
-    }
-    this._domainEventBus.raise({
-      eventName: EVENT_STATUS,
-      payload: {
-        TCData
-      }
+    this._displayStatusRepository.setDisplayStatus({
+      newStatus: visible ? DisplayStatus.VISIBLE : DisplayStatus.HIDDEN
     })
+    this._eventStatusService.updateUiStatus()
   }
 }
